+++ +gamrywb/+util/csvEscape.m
@@ -1,5 +1,0 @@
-function s = csvEscape(x)
-%CSVESCAPE Escape double quotes for CSV text fields.
-
-    s = strrep(char(x), '"', '""');
-end